rename checkers back to loaders
